Extract IgniteLockFailoverTests to a separate file to avoid side effects on other tests
